# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Backend API for the Share Market Simulator with high-sensitivity price engine and order matching.",
        "title": "Share Market Simulator API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "2.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/admin/companies": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Register a new company in the system",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Create a company (Admin only)",
                "parameters": [
                    {
                        "description": "Company details",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_ipo.CreateCompanyRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/admin/ipos": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Start an IPO for a specific company",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Launch an IPO (Admin only)",
                "parameters": [
                    {
                        "description": "IPO details",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_ipo.LaunchIPORequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/admin/ipos/{id}/allocate": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Trigger the lottery-based allocation for a closed IPO",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Allocate IPO shares (Admin only)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "IPO ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
        "/ipos": {
            "get": {
                "description": "Get a list of all IPOs (active, closed, allocated)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ipo"
                ],
                "summary": "List all IPOs",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/ipos/{id}": {
            "get": {
                "description": "Get details of a specific IPO by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ipo"
                ],
                "summary": "Get IPO details",
                "parameters": [
                    {
                        "type": "string",
                        "description": "IPO ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/ipos/{id}/apply": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Apply for shares in an open IPO",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ipo"
                ],
                "summary": "Apply for IPO",
                "parameters": [
                    {
                        "type": "string",
                        "description": "IPO ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Application details",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_ipo.ApplyIPORequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "internal_ipo.ApplyIPORequest": {
            "type": "object",
            "required": [
                "shares_requested"
            ],
            "properties": {
                "shares_requested": {
                    "type": "integer"
                }
            }
        },
        "internal_ipo.CreateCompanyRequest": {
            "type": "object",
            "required": [
                "name",
                "symbol",
                "total_supply"
            ],
            "properties": {
                "name": {
                    "type": "string"
                },
                "sector": {
                    "type": "string"
                },
                "symbol": {
                    "type": "string"
                },
                "total_supply": {
                    "type": "integer"
                }
            }
        },
        "internal_ipo.LaunchIPORequest": {
            "type": "object",
            "required": [
                "close_at",
                "company_id",
                "max_per_applicant",
                "open_at",
                "price_per_share",
                "total_shares"
            ],
            "properties": {
                "close_at": {
                    "type": "string"
                },
                "company_id": {
                    "type": "string"
                },
                "max_per_applicant": {
                    "type": "integer"
                },
                "open_at": {
                    "type": "string"
                },
                "price_per_share": {
                    "type": "string"
                },
                "total_shares": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and then your token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}